requirements.txt
# Core Dependencies
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1

# Data Processing
pandas==2.2.0
numpy==1.26.4
plotly==5.18.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.0.1
pytesseract==0.3.10
pdf2image==1.17.0

# Audio Processing
speechrecognition==3.10.1
pydub==0.25.1

# UI Components
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.12

# API and Networking
requests==2.31.0
aiohttp==3.9.3

# Vector Database (for knowledge base)
chromadb==0.4.22
sentence-transformers==2.5.1

# Utilities
tqdm==4.66.2